+++ src/main/kotlin/com/example/team/haribo/goms/gomsserverv3/global/exception/ErrorCode.kt
@@ -1,4 +1,4 @@
-package com.example.team.haribo.goms.GomsServerV3.global.exception
+package com.example.team.haribo.goms.gomsserverv3.global.exception
 
 enum class ErrorCode(
     val status: Int,
