+++ src/main/kotlin/com/example/team/haribo/goms/gomsserverv3/global/exception/ErrorResponse.kt
@@ -1,4 +1,4 @@
-package com.example.team.haribo.goms.GomsServerV3.global.exception
+package com.example.team.haribo.goms.gomsserverv3.global.exception
 
 data class ErrorResponse(
     val status: Int,
