+++ src/main/kotlin/com/example/team/haribo/goms/gomsserverv3/global/jwt/JwtProperties.kt
@@ -1,4 +1,4 @@
-package com.example.team.haribo.goms.GomsServerV3.global.jwt
+package com.example.team.haribo.goms.gomsserverv3.global.jwt
 
 import org.springframework.boot.context.properties.ConfigurationProperties
