+++ src/test/kotlin/com/example/team/haribo/goms/gomsserverv3/GomsServerV3ApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.example.team.haribo.goms.gomsserver_v3
+package com.example.team.haribo.goms.gomsserverv3
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
